Rename MySQL::DatabaseStatements => Mysql2::DatabaseStatements
